+++ CHANGES.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.2.1] - 2025-04-23
+* Added fuzzy date parsing
+* Changed script name to `timex-app`
+
 ## [0.2.0] - 2024-09-24
 * Added calculation mode